% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_diag_pairs_for_all.R
\name{add_top_center_label}
\alias{add_top_center_label}
\title{Add a single centered label at the top of a panel (transparent overlay)}
\usage{
add_top_center_label(p, label, size = 10, color = "red")
}
\description{
Add a single centered label at the top of a panel (transparent overlay)
}
\keyword{internal}
