% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R, R/utility.R
\name{calc.gamma}
\alias{calc.gamma}
\title{Calculate gamma from the Pella-Tomlinson exponent}
\usage{
calc.gamma(n)

calc.gamma(n)
}
\arguments{
\item{n}{Exponent of the Pella-Tomlinson surplus production equation.}
}
\value{
A numeric value of gamma.
}
\description{
This function computes the scaling factor gamma used in the Pella-Tomlinson surplus production function.
}
\examples{
calc.gamma(2)

}
