% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.inp.R
\name{check.inp}
\alias{check.inp}
\title{Check list of input variables}
\usage{
check.inp(inp, verbose = TRUE, mancheck = TRUE)
}
\arguments{
\item{inp}{List of input variables. See Details for required variables.}

\item{verbose}{Should detailed outputs be provided (default: TRUE).}

\item{mancheck}{Should the time-dependent objects in \code{inp} be checked against the management time and corrected if necessary? (Default: TRUE)}
}
\value{
The checked input list, possibly modified or with missing values filled.
}
\description{
Fills in default values if missing. Optionally checks time-dependent objects against the management time and corrects if necessary.
}
\details{
This function checks the input list for the ELU model, fills in missing default values, and can check time-dependent objects against the management time.
}
