% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_param_pct_change.R
\name{compute_param_pct_change}
\alias{compute_param_pct_change}
\title{Compute \% changes vs base scenario (default "S4")}
\usage{
compute_param_pct_change(
  param_df,
  base_scenario = "S4",
  keep_ratio_year = 2022,
  include_ci = FALSE,
  families_order = c("Pella", "Fox", "Schaefer"),
  params_order = c("K", "r", "MSY", "B/Bmsy", "F/Fmsy")
)
}
\arguments{
\item{param_df}{...}

\item{base_scenario}{...}

\item{keep_ratio_year}{...}

\item{include_ci}{...}

\item{families_order}{...}

\item{params_order}{...}
}
\value{
...
}
\description{
Compute \% changes vs base scenario (default "S4")
}
