% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_diag_pairs_for_all.R
\name{.get_diag_fun}
\alias{.get_diag_fun}
\title{Locate the diagnostic plotting function you already have}
\usage{
.get_diag_fun()
}
\description{
Locate the diagnostic plotting function you already have
}
\keyword{internal}
