% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_osar_diag_pairs_for_all.R
\name{.get_osar_diag_fun}
\alias{.get_osar_diag_fun}
\title{Locate the OSAR diagnostic plotting function}
\usage{
.get_osar_diag_fun()
}
\description{
Locate the OSAR diagnostic plotting function
}
\keyword{internal}
