% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_osar_diag_pairs_for_all.R
\name{ensure_osar_resid}
\alias{ensure_osar_resid}
\title{Ensure OSAR residuals exist on a fitted object}
\usage{
ensure_osar_resid(fit)
}
\description{
Ensure OSAR residuals exist on a fitted object
}
\keyword{internal}
