% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_diag_pairs_for_all.R
\name{ensure_process_resid}
\alias{ensure_process_resid}
\title{Ensure process residuals exist on a fitted object}
\usage{
ensure_process_resid(fit)
}
\description{
Ensure process residuals exist on a fitted object
}
\keyword{internal}
