% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_prior_table_tex.R
\name{export_prior_table_tex}
\alias{export_prior_table_tex}
\title{Export Prior Table for a Scenario to LaTeX}
\usage{
export_prior_table_tex(
  S_Fox,
  S_Schaefer,
  S_Pella,
  scenario_name = "1",
  output_dir = "."
)
}
\arguments{
\item{S_Fox}{A SPiCT model fit for the Fox model.}

\item{S_Schaefer}{A SPiCT model fit for the Schaefer model.}

\item{S_Pella}{A SPiCT model fit for the Pella-Tomlinson model.}

\item{scenario_name}{A string used to name the output file (e.g., \code{"1"} or \code{"S1"}).}

\item{output_dir}{A path to the folder where the .tex file should be saved. Default is current directory.}
}
\value{
Invisibly returns the LaTeX string for further use if needed.
}
\description{
This function extracts the priors from SPiCT fits (Fox, Schaefer, Pella),
formats them in LaTeX, and saves a \code{table_scenarioX.tex} file for Beamer inclusion.
}
