% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.elu2.R
\name{fit.elu2}
\alias{fit.elu2}
\title{Fit elu2 Model}
\usage{
fit.elu2(inp, verbose = TRUE, dbg = 0)
}
\arguments{
\item{inp}{List of input variables as output by check.inp.}

\item{verbose}{Should detailed outputs be provided (default: TRUE).}

\item{dbg}{Debugging option. Will print out runtime information useful for debugging if set to 1. Will print even more if set to 2.}
}
\value{
A result report containing estimates of model parameters, random effects (biomass and fishing mortality), reference points (Fmsy, Bmsy, MSY) including uncertainties given as standard deviations.
}
\description{
Fits the customized ELU2 stock assessment model using TMB.
}
