% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-exports.R, R/data-inp1.glm.R
\docType{data}
\name{inp1.glm}
\alias{inp1.glm}
\title{SPiCT Input List from Delta GLM}
\format{
A list.

A named list suitable for use in \code{fit.spict()}.
}
\usage{
inp1.glm

inp1.glm
}
\description{
inp1.glm

A \code{list} formatted input object for SPiCT model fitting, constructed using delta-GLM standardized indices.
}
\details{
This input object includes \code{obsC}, \code{timeC}, \code{obsI}, and \code{timeI} derived from delta-GLM indices.
}
\seealso{
\link{Delta_GLM_list}, \link{long_series_list}
}
\keyword{datasets}
