% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-exports.R, R/data-inp1.sdmTMB.R
\docType{data}
\name{inp1.sdmTMB}
\alias{inp1.sdmTMB}
\title{SPiCT Input List from sdmTMB}
\format{
A list.

A \code{list} containing the required fields for \code{fit.spict()}.
}
\usage{
inp1.sdmTMB

inp1.sdmTMB
}
\description{
inp1.sdmTMB

A SPiCT input object using standardized indices derived from \code{sdmTMB} model predictions.
}
\details{
Typically includes one or more short time series, used in scenario comparisons.
}
\seealso{
\link{Long_sdmTMB_list}, \link{short_sdmTMB_list}
}
\keyword{datasets}
