% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retro.rho.ndiaye2.R
\docType{data}
\name{jabba_colors}
\alias{jabba_colors}
\title{JABBA Color Set for Peel Plots}
\format{
A character vector of 8 hexadecimal color codes.
}
\usage{
jabba_colors
}
\description{
Predefined color palette inspired by the JABBA package, suitable for use in
retrospective "peel" plots or any situation requiring distinct colors.
}
\details{
The colors are: red, blue, green, purple, orange, brown, pink, and grey.
}
\examples{
barplot(rep(1, 8), col = jabba_colors)

}
\keyword{datasets}
