% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Long_delta_GLM_list.R
\docType{data}
\name{Long_delta_GLM_list}
\alias{Long_delta_GLM_list}
\title{Long-Series Delta GLM Indices}
\format{
A list of numeric vectors with standardized CPUE values over time.
}
\usage{
Long_delta_GLM_list
}
\description{
A list of delta-GLM standardized indices representing the long time series (e.g., 1992–2016).
}
\details{
Used to construct long \code{obsI} time series for input to SPiCT.
}
\seealso{
\link{Delta_GLM_list}, \link{inp1.glm}
}
\keyword{datasets}
