% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Long_sdmTMB_list.R, R/data-exports.R
\docType{data}
\name{Long_sdmTMB_list}
\alias{Long_sdmTMB_list}
\title{Long-Series sdmTMB Indices}
\format{
A named list of numeric vectors with standardized values per year.

A named list.
}
\usage{
Long_sdmTMB_list

Long_sdmTMB_list
}
\description{
A list of long standardized CPUE indices generated using \code{sdmTMB} model outputs.

Long_sdmTMB_list
}
\details{
Used to construct longer time series in sdmTMB-based SPiCT input objects.
}
\seealso{
\link{inp1.sdmTMB}
}
\keyword{datasets}
