% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-exports.R, R/data-long_series_list.R
\docType{data}
\name{long_series_list}
\alias{long_series_list}
\title{Long Series CPUE Indices}
\format{
A named list.

A named list of numeric vectors, one per survey or season.
}
\usage{
long_series_list

long_series_list
}
\description{
long_series_list

A list of standardized long CPUE indices used in management scenario simulations.
}
\details{
Used to populate the long-term \code{obsI} slot in SPiCT input structures.
}
\seealso{
\link{inp1.glm}
}
\keyword{datasets}
