% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_plot_manage_ffmsy_panel_sim.R
\name{man_cols}
\alias{man_cols}
\title{Repeatable colour palette for management scenarios}
\usage{
man_cols(n)
}
\arguments{
\item{n}{Integer. Number of colours required.}
}
\value{
A character vector of length \code{n} with colour names.
}
\description{
Returns a repeatable vector of colours to map across an arbitrary number
of management scenarios. Colours recycle to the requested length.
}
\examples{
if (interactive()) man_cols(5)

}
\keyword{internal}
