% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_plot_manage_bbmsy_panel.R
\name{my_plot_manage_bbmsy_panel}
\alias{my_plot_manage_bbmsy_panel}
\title{B/Bmsy panel (ggplot2) with optional scenario overlays}
\usage{
my_plot_manage_bbmsy_panel(
  rep_man,
  scenario_color = NULL,
  show_CIs = TRUE,
  CI = 0.95,
  show_legend = TRUE
)
}
\description{
B/Bmsy panel (ggplot2) with optional scenario overlays
}
