% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_plot_manage_f_panel.R
\name{my_plot_manage_f_panel}
\alias{my_plot_manage_f_panel}
\title{Absolute F\link{t} panel (with Fmsy band, (F/Fmsy)×Fmsy ribbon, scenarios)}
\usage{
my_plot_manage_f_panel(
  rep_man,
  scenario_color = NULL,
  show_CIs = TRUE,
  CI = 0.95,
  show_legend = FALSE
)
}
\description{
Absolute F\link{t} panel (with Fmsy band, (F/Fmsy)×Fmsy ribbon, scenarios)
}
