% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_retro_grid.ELU4_single.R
\name{plot_retro_grid.ELU4_single}
\alias{plot_retro_grid.ELU4_single}
\title{Plot individual retrospective grids for each SPiCT model in 2x2 layout}
\usage{
plot_retro_grid.ELU4_single(
  models,
  width = 8,
  height = 5,
  dpi = 400,
  output_dir = NULL
)
}
\arguments{
\item{models}{A named list of SPiCT model fits (each with $retro)}

\item{width, }{height, dpi Figure size and resolution}

\item{output_dir}{If given, saves plots to this folder}
}
\value{
Invisibly returns a list of patchwork plots
}
\description{
For each model in the list, generates and optionally saves a 2x2 retrospective grid:
Top-left: B\link{t}, Bottom-left: B\link{t}/B\link{MSY}, Top-right: F\link{t}, Bottom-right: F\link{t}/F\link{MSY}
}
