% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotspict.diagnostic.process_gg.R
\name{plotspict.diagnostic.process_gg}
\alias{plotspict.diagnostic.process_gg}
\title{Process diagnostics (ggplot2; blue points + light-blue lines on top 4 panels)}
\usage{
plotspict.diagnostic.process_gg(
  rep,
  lag.max = 4,
  qlegend = TRUE,
  plot.data = TRUE,
  mfcol = FALSE,
  add.loess = FALSE,
  span = 0.75,
  stamp = get.version()
)
}
\description{
Process diagnostics (ggplot2; blue points + light-blue lines on top 4 panels)
}
