% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotspict.priors.ggplot99.R
\name{plotspict.priors.ggplot99}
\alias{plotspict.priors.ggplot99}
\title{Plot priors and posterior distributions using ggplot2 + theme_sleek}
\usage{
plotspict.priors.ggplot99(
  rep,
  do.plot = NULL,
  stamp = if (exists("get.version")) get.version() else "v0.0",
  CI = 0.95
)
}
\arguments{
\item{rep}{A result from fit.spict(), or a list with \verb{$result}}

\item{do.plot}{Maximum number of priors to display (NULL = all)}

\item{stamp}{Optional version label}

\item{CI}{Posterior confidence level}
}
\value{
ggplot object (invisibly)
}
\description{
Plot priors and posterior distributions using ggplot2 + theme_sleek
}
