% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_kobe_everything.R
\name{save_kobe_everything}
\alias{save_kobe_everything}
\title{Save individuals and grids together (one-liner helper)}
\usage{
save_kobe_everything(
  all_models,
  out_root_individuals = file.path("FIG", "KobeIndividuals"),
  out_root_grids = file.path("FIG", "KobePhases"),
  ...
)
}
\description{
Save individuals and grids together (one-liner helper)
}
