% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-exports.R
\docType{data}
\name{scenario_definitions}
\alias{scenario_definitions}
\title{Scenario definitions for SPiCT modeling}
\format{
A named list.
}
\usage{
scenario_definitions
}
\description{
scenario_definitions
}
\keyword{datasets}
