% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-exports.R, R/data-short_sdmTMB_list.R
\docType{data}
\name{short_sdmTMB_list}
\alias{short_sdmTMB_list}
\title{Short-Series sdmTMB Indices}
\format{
A named list.

A named list of numeric vectors representing recent survey periods.
}
\usage{
short_sdmTMB_list

short_sdmTMB_list
}
\description{
short_sdmTMB_list

A list of short CPUE indices (e.g., 2017–2022) standardized using the \code{sdmTMB} delta model.
}
\details{
Typically used to simulate data-limited SPiCT scenarios or to compare with delta-GLM results.
}
\seealso{
\link{inp1.sdmTMB}, \link{Long_sdmTMB_list}
}
\keyword{datasets}
