% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.elu6.R
\name{smart_label}
\alias{smart_label}
\title{Smart Label Formatter for Axis Ticks (with Scientific Notation for q1/q2)}
\usage{
smart_label(x)
}
\arguments{
\item{x}{Numeric vector of axis tick values.}
}
\value{
A character vector of formatted labels.
}
\description{
Converts numeric values into readable labels. Uses scientific notation if values have more than two decimal digits or are very small/large.
}
