% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotspict.diagnostic.process_gg.R
\name{theme_minimal_compact2_diag}
\alias{theme_minimal_compact2_diag}
\title{Compact diagnostic theme (no frills, bold panels, no legends)}
\usage{
theme_minimal_compact2_diag(base_size = 10, base_family = "")
}
\arguments{
\item{base_size}{Base font size.}

\item{base_family}{Base font family.}
}
\value{
A ggplot2 theme.
}
\description{
Compact diagnostic theme (no frills, bold panels, no legends)
}
