% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotspict.priors.ggplot98.R
\name{theme_sleek}
\alias{theme_sleek}
\title{A sleek, light–based theme for ggplot2}
\usage{
theme_sleek(base_size = 11, base_family = "")
}
\arguments{
\item{base_size}{Base font size}

\item{base_family}{Base font family}
}
\description{
A sleek, light–based theme for ggplot2
}
