% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{calc.gamma}
\alias{calc.gamma}
\title{Calculate gamma from the Pella-Tomlinson exponent}
\usage{
calc.gamma(n)
}
\arguments{
\item{n}{Numeric; exponent of the Pella-Tomlinson production function.}
}
\value{
A numeric value of gamma.
}
\description{
This function computes the scaling factor gamma used in the Pella-Tomlinson surplus production function.
}
\examples{
calc.gamma(2)

}
