% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Delta_GLM_list.R, R/data-exports.R
\docType{data}
\name{Delta_GLM_list}
\alias{Delta_GLM_list}
\title{Delta GLM Model List}
\format{
A named list of fitted model objects. Each element corresponds to a year-season combination.
The models are generally of class \code{glm} or \code{lm}, depending on the family used.

A named list of model outputs.
}
\usage{
Delta_GLM_list

Delta_GLM_list
}
\description{
A list of delta-GLM fitted models used for shrimp CPUE standardization.

Delta_GLM_list
}
\details{
This object was created during index standardization prior to inclusion in the SPiCT model.
It is used to extract time series of standardized CPUE for short/long series comparison.
}
\seealso{
\link{inp1.glm}, \link{Long_delta_GLM_list}
}
\keyword{datasets}
